% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{brier_score}
\alias{brier_score}
\title{Brier Score function}
\usage{
brier_score(prob, observed, levels)
}
\arguments{
\item{prob}{predicted probabilities}

\item{observed}{$y$ observed values (it assumed that the positive class is coded is equal to one and the negative 0)}

\item{levels}{A string vector with the original levels from the target variable}
}
\description{
Brier Score function
}
