% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{predict.rm_class}
\alias{predict.rm_class}
\title{Prediction function for the rm_class_model}
\arguments{
\item{object}{A fitted RM model object of class  \code{rm_class}.}

\item{newdata}{A data frame or matrix containing the new data to be predicted.}

\item{...}{currently not used.}
}
\value{
A vector of predicted outcomes: probabilities in case of `prob_model = TRUE` and classes in case of `prob_model = FALSE`.
}
\description{
This function predicts the outcome for a RM object model using new data
}
\examples{
NULL
}
