% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{predict.rm_model_class}
\alias{predict.rm_model_class}
\title{Prediction function for the rm_class_model}
\usage{
predict.rm_model_class(mod, newdata)
}
\arguments{
\item{mod}{A fitted RM model object of class "rm_model"}

\item{newdata}{A data frame or matrix containing the new data to be predicted}
}
\value{
A vector of predicted outcomes: probabilities in case of `prob_model = TRUE` and classes in case of `prob_model = FALSE`
}
\description{
This function predicts the outcome for a RM object model using new data
}
\examples{
library(rmachines)
sim_data <- rmachines::sim_class(n = 100)
rm_mod <- rmachines::random_machines(y~., train = sim_data, valiation = sim_data)
rm_pred <- predict(rm_mod, newdata = sim_data)

}
