% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{predict.rm_model_reg}
\alias{predict.rm_model_reg}
\title{Prediction function for the rm_reg_model}
\usage{
\method{predict}{rm_model_reg}(mod, newdata)
}
\arguments{
\item{mod}{A fitted RM model object of reg "rm_model"}

\item{newdata}{A data frame or matrix containing the new data to be predicted}
}
\value{
A vector of predicted outcomes: probabilities in case of `prob_model = TRUE` and reges in case of `prob_model = FALSE`
}
\description{
This function predicts the outcome for a RM object model using new data
}
\examples{
library(rmachines)
sim_data <- rmachines::sim_reg(n = 100)
rm_mod <- rmachines::random_machines(y~., train = sim_data, validation = sim_data)

}
