% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{predict.rm_reg}
\alias{predict.rm_reg}
\alias{predict.rm_reg,}
\alias{predict}
\title{Predict method for rm_reg class}
\arguments{
\item{object}{An object of class "rm_reg"}

\item{newdata}{New data for prediction}

\item{...}{other arguments}
}
\value{
Predicted values
}
\description{
Predict method for rm_reg class
}
