% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\docType{class}
\name{rm_reg-class}
\alias{rm_reg-class}
\alias{rm_reg}
\title{S4 class for RM regression}
\description{
S4 class for RM regression
}
\details{
For more details see Ara, Anderson, et al. "Regression random machines: An ensemble support vector regression model with free kernel choice." Expert Systems with Applications 202 (2022): 117107.
}
\section{Slots}{

\describe{
\item{\code{y_train_hat}}{a numeric  corresponding to the predictions \eqn{\hat{y}_{i}} for the training set}

\item{\code{lambda_values}}{a named list with value of the vector of \eqn{\boldsymbol{\lambda}} sampling probabilities associated with each each kernel function}

\item{\code{model_params}}{a list with all used model specifications}

\item{\code{bootstrap_models}}{a list with all \code{ksvm} objects for each bootstrap sample}

\item{\code{bootstrap_samples}}{a list with all bootstrap samples used to train each base model of the ensemble}

\item{\code{kernel_weight_norm}}{a numeric vector corresponding to the normalised weights for each bootstrap model contribution}
}}

