% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{RMSE}
\alias{RMSE}
\title{#Root Mean Squared Error (RMSE) Function}
\usage{
RMSE(predicted, observed)
}
\arguments{
\item{predicted}{A vector of predicted values \eqn{\hat{\mathbf{y}}}.}

\item{observed}{A vector of observed values \eqn{\mathbf{y}}.}
}
\value{
a the Root Mean Squared error calculated by \eqn{RMSE= \frac{1}{n}\sum_{i=1}^{n}\sqrt{\left(\hat{y}_{i}-\y_{i}\right)^{2}}}
}
\description{
#Root Mean Squared Error (RMSE) Function
}
