% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_examples.R
\name{sim_reg}
\alias{sim_reg}
\title{Simple regression \eqn{y = x_{1}^{2} + e^{x_{2}^{2}} + \varepsilon} case based on Ara et. al 2022}
\usage{
sim_reg(n, seed = NULL)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Define a seed to run the simulation. NULL is the default}
}
\value{
A simulated data.frame with two predictors and the target variable.
}
\description{
Simple regression \eqn{y = x_{1}^{2} + e^{x_{2}^{2}} + \varepsilon} case based on Ara et. al 2022
}
\examples{
library(rmachines)
sim_data <- sim_reg(n=100)
}
