% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{random_machines}
\alias{random_machines}
\title{Call the Random Machines model for a regression or classification task.}
\usage{
random_machines(
  formula,
  train,
  validation,
  boots_size = 25,
  cost = 10,
  seed.bootstrap = NULL,
  automatic_tuning = FALSE,
  gamma_rbf = 1,
  gamma_lap = 1,
  degree = 2,
  poly_scale = 1,
  offset = 0,
  gamma_cau = 1,
  d_t = 2,
  kernels = c("rbfdot", "polydot", "laplacedot", "vanilladot", "cauchydot"),
  prob_model = T,
  loss_function = RMSE,
  epsilon = 0.1,
  beta = 2
)
}
\arguments{
\item{formula}{Insert the formula of the model}

\item{train}{the training data $\left\{\left( \mathbf{x}_{i},y_{i} \right)  \right\}_{i=1}^{N} used to train the model}

\item{validation}{the validation data $\left\{\left( \mathbf{x}_{i},y_{i} \right)  \right\}_{i=1}^{V} used to calculate probabilities $\lambda_{r}$}

\item{boots_size}{number of bootstrap samples}

\item{cost}{it is the "C" constant of the regularisation of the SVM Lagrange formulation}

\item{seed.bootstrap}{setting a seed to replicate bootstrap sampling. The default value is $\texttt{NULL}$}

\item{automatic_tuning}{boolean to define if the kernel hyperparameters will be selected using the $\texttt{sigest}$ from the $\texttt{ksvm}$ function}

\item{gamma_rbf}{the hyperparameter $\gamma_{RBF}$ used in the RBF kernel}

\item{gamma_lap}{the hyperparameter $\gamma_{LAP}$ used in the Laplacian kernel}

\item{degree}{the degree of used in the Polynomial kernel}

\item{poly_scale}{the scale parameter from Polynomial kernel}

\item{offset}{the offset parameter from the Polynomial kernel}

\item{gamma_cau}{the offset parameter from the Cauchy kernel}

\item{d_t}{??}

\item{kernels}{the vector with the kernel functions that will me used in the random machines.}

\item{prob_model}{a boolean to define if the algorithm will be using a probabilistic approach to the define the predictions (default = $\texttt{R})}

\item{loss_function}{Define which loss function is gonna be used}

\item{epsilon}{The epsilon in the loss function used from the SVR implementation.}

\item{beta}{The correlation parameter $\beta$ which calibrate the penalisation of each kernel performance.}
}
\description{
Call the Random Machines model for a regression or classification task.
}
