% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_examples.R
\name{sim_class}
\alias{sim_class}
\title{Function to generate a two classification data set from normal distribution}
\usage{
sim_class(
  n,
  p = 2,
  ratio = 0.5,
  mean_one = 0,
  sd_one = 1,
  mean_two = 1,
  sd_two = 1,
  seed = NULL
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of predictors}

\item{ratio}{Ratio between class A and class B}

\item{mean_one}{Mean of \eqn{X_{1}}.}

\item{sd_one}{Standard deviation of \eqn{X_{1}}.}

\item{mean_two}{Mean of \eqn{X_{2}}}

\item{sd_two}{Standard devation of \eqn{X_{2}}}

\item{seed}{Setting a seed for reproducibility of results. The default is NULL}
}
\value{
A simulated data.frame with two predictors for a binary classification problem
}
\description{
This function predicts the outcome for a RM object model using new data
}
\examples{
library(rmachines)
sim_data <- sim_class(n = 100)
}
